feat: base de données configurable, et portable au-delà de SQLite
`ZANSHIN_DATABASE_URL` existait et était documentée comme LE moyen de pointer
Zanshin vers une autre base. Le moteur passait `check_same_thread` — un argument
propre à SQLite — à tous les dialectes, donc la variable était réelle pour un
dialecte et un piège pour les autres. Détail en §9decies de l'ADR.

Configuration
- Deux boutons : ZANSHIN_DB_PATH (chemin de fichier, rendu absolu parce qu'un
  chemin relatif se résout différemment selon reflex/alembic/systemd) et
  ZANSHIN_DATABASE_URL (URL complète, prioritaire).
- Options de moteur par dialecte : contrôle inter-threads et délai d'attente sur
  verrou pour SQLite ; pool et pool_pre_ping pour une base serveur.
- Dialecte inconnu, URL illisible, pilote absent : trois erreurs distinctes, à
  l'import, avec un message qui nomme le problème.
- ZANSHIN_AUTO_MIGRATE=false pour migrer comme étape de déploiement. Sauter
  n'est pas ignorer : le schéma est vérifié, une base en retard arrête le
  démarrage.
- Le verrou de migration déménage à côté de la base : c'est l'accès à cette base
  qui est sérialisé, pas l'arborescence source.

Portabilité — six défauts trouvés en exécutant, aucun visible depuis SQLite
- GUID rendait `BINARY`, type inexistant sur PostgreSQL : implémentation par
  dialecte (uuid natif là-bas, 16 octets ailleurs, stockage SQLite inchangé).
- Les migrations 0003 et 0004 faisaient `FROM user` en SQL brut, mot réservé
  PostgreSQL où il désigne la fonction current_user.
- SafeDateTime déclarait String sans longueur : MySQL refuse VARCHAR sans
  longueur.
- Toutes les clés étrangères étaient BigInteger vers des clés primaires Integer.
  Corrigé en rétrécissant les FK, pas en élargissant les PK : sur SQLite seule
  une colonne `INTEGER PRIMARY KEY` est un alias du rowid. Migration 0007 pour
  l'existant, 0001/0002/0003 amendées pour qu'une base neuve soit correcte.
- 0003 supprimait une colonne sans supprimer d'abord sa contrainte de clé
  étrangère, dont le nom est cherché par inspection puisqu'il varie.
- DROP INDEX IF EXISTS et NULLS LAST n'existent pas sur MySQL — le second dans
  le tri applicatif de la liste des problèmes, remplacé par `colonne IS NULL`.

Tests
- tests/test_database_backends.py : 36 tests via testcontainers sur PostgreSQL
  16 et MySQL 8.4 — les sept migrations, alembic check, puis chaque type de
  colonne maison et chaque service qui en possède un. Marqueur `backends`,
  exclus par défaut, sautés sans Docker.
- tests/test_database_config.py : précédence des variables, options par
  dialecte, échecs lisibles, mot de passe absent des journaux.
- Deux défauts de mon propre code trouvés là : les arguments par défaut figeaient
  DATABASE_URL à l'import, et restrict_database_permissions n'était pas
  testable sans toucher aux globales d'un autre module.

623 tests sur SQLite, 36 sur PostgreSQL et MySQL. La base de développement passe
à 0007 sans perte (429 problèmes, 1258 findings, 421 directivités conservées) et
l'application démarre sur les deux backends.

Correction d'une vérification annoncée à tort : les contrôles de migration des
deux commits précédents utilisaient ZANSHIN_DB_PATH, qui n'existait pas encore,
donc les commandes alembic retombaient sur la base réelle — aucune base neuve
n'avait été testée. Refaits ici avec la variable qui existe désormais.

+++ migrations/versions/0002_issue_tracking_and_actionable_finding_data.py
@@ -38,11 +38,14 @@ depends_on: Union[str, Sequence[str], None] = None
 
 
 def upgrade() -> None:
+    # Note: the foreign key columns created here were originally `BigInteger` against
+    # `Integer` primary keys, which MySQL refuses. Amended so a fresh database can be
+    # built on any backend; databases created by the original are corrected by 0007.
     op.create_table(
         "issue",
         sa.Column("id", sa.Integer(), nullable=False),
-        sa.Column("repo_id", sa.BigInteger(), nullable=True),
-        sa.Column("container_id", sa.BigInteger(), nullable=True),
+        sa.Column("repo_id", sa.Integer(), nullable=True),
+        sa.Column("container_id", sa.Integer(), nullable=True),
         sa.Column("fingerprint", sa.String(length=64), nullable=False),
         sa.Column("type", sa.String(length=50), nullable=False),
         sa.Column("identifier", sa.String(length=255), nullable=True),
@@ -64,8 +67,8 @@ def upgrade() -> None:
         sa.Column("first_seen_at", zanshin.models.safedatetime.SafeDateTime(), nullable=False),
         sa.Column("last_seen_at", zanshin.models.safedatetime.SafeDateTime(), nullable=False),
         sa.Column("resolved_at", zanshin.models.safedatetime.SafeDateTime(), nullable=True),
-        sa.Column("first_seen_scan_id", sa.BigInteger(), nullable=True),
-        sa.Column("last_seen_scan_id", sa.BigInteger(), nullable=True),
+        sa.Column("first_seen_scan_id", sa.Integer(), nullable=True),
+        sa.Column("last_seen_scan_id", sa.Integer(), nullable=True),
         sa.Column("times_seen", sa.Integer(), nullable=False, server_default="1"),
         sa.Column("triage_status", sa.String(length=30), nullable=False, server_default="under_review"),
         sa.Column("triage_justification", sa.String(length=64), nullable=True),
